api_gateway/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
httpx
python-dotenv
pydantic
websockets
python-multipart

